Вынесение в отдельные файлы классы. Добавление меню "О программе"
