+++ Lesson08/Task03/Program.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using System.Xml.Serialization;
 
 namespace Task03
 {
@@ -29,142 +26,4 @@ namespace Task03
 			Application.Run(new frmMain());
 		}
 	}
-
-	[Serializable]
-	public class Question
-	{
-		public string text;       // Текст вопроса
-		public bool trueFalse;// Правда или нет
-													// Здесь мы нарушаем правила инкапсуляции и эти поля нужно было бы реализовать через открытые свойства, но для упрощения примера оставим так
-													// Вам же предлагается сделать поля закрытыми и реализовать открытые свойства Text и TrueFalse
-													// Для сериализации должен быть пустой конструктор.
-		public Question()
-		{
-		}
-		public Question(string text, bool trueFalse)
-		{
-			this.text = text;
-			this.trueFalse = trueFalse;
-		}
-	}
-	// Класс для хранения списка вопросов. А также для сериализации в XML и десериализации из XML
-	class TrueFalse
-	{
-		string fileName;
-		List<Question> list;
-		public string FileName
-		{
-			get { return fileName; }
-			set { fileName = value; }
-		}
-		public TrueFalse(string fileName)
-		{
-			this.fileName = fileName;
-			list = new List<Question>();
-		}
-		public void Add(string text, bool trueFalse)
-		{
-			list.Add(new Question(text, trueFalse));
-		}
-		public void Remove(int index)
-		{
-			if (list != null && index < list.Count && index >= 0) list.RemoveAt(index);
-		}
-		// Индексатор - свойство для доступа к закрытому объекту
-		public Question this[int index]
-		{
-			get { return list[index]; }
-		}
-		public void Save()
-		{
-			XmlSerializer xmlFormat = new XmlSerializer(typeof(List<Question>));
-			Stream fStream = new FileStream(fileName, FileMode.Create, FileAccess.Write);
-			xmlFormat.Serialize(fStream, list);
-			fStream.Close();
-		}
-		public void Load()
-		{
-			XmlSerializer xmlFormat = new XmlSerializer(typeof(List<Question>));
-			Stream fStream = new FileStream(fileName, FileMode.Open, FileAccess.Read);
-			list = (List<Question>)xmlFormat.Deserialize(fStream);
-			fStream.Close();
-		}
-		public int Count
-		{
-			get { return list.Count; }
-		}
-	}
-
-
-	/*
-	/// <summary>
-	/// Класс для хранения объекта вопроса
-	/// </summary>
-	[Serializable]
-	public class Question
-	{
-		private string _text;
-		private bool _amswer;
-		public Question() { }
-		public Question(string QuestionText, bool Answer)
-		{
-			_text = QuestionText;
-			_amswer = Answer;
-		}
-		public string QuestionText { get { return _text; } }
-		public bool CorrectAnswer { get { return _amswer; } }
-	}
-
-	public class ListQuestion
-	{
-		string _fileName;
-		List<Question> _list;
-		public ListQuestion(string FileName)
-		{
-			_fileName = FileName;
-			_list = new List<Question>();
-		}
-		public string FileName { get { return _fileName; } set { _fileName = value; } }
-		public void Add(string QuestionText, bool CorrectAnswer)
-		{ _list.Add(new Question(QuestionText, CorrectAnswer)); }
-		public bool Remove(int Index)
-		{
-			bool res = false;
-			if (_list != null && Index < _list.Count && Index >= 0)
-			{
-				res = true;
-				_list.RemoveAt(Index);
-			}
-			return res;
-		}
-		public Question this[int Index] { get { return _list[Index]; } }
-		public int Count { get { return _list.Count; } }
-		public void Load()
-		{
-			XmlSerializer xmlSer = new XmlSerializer(typeof(List<Question>));
-			Stream fs = new FileStream(_fileName, FileMode.Open, FileAccess.Read);
-			_list = (List<Question>)xmlSer.Deserialize(fs);
-			fs.Close();
-		}
-		public bool Save()
-		{
-			bool res = false;
-			XmlSerializer xmlSer = new XmlSerializer(typeof(List<Question>));
-			try
-			{
-				using (Stream fs = new FileStream(_fileName, FileMode.Create, FileAccess.Write))
-				{
-					xmlSer.Serialize(fs, _list);
-				}
-			}
-			catch (Exception ex)
-			{
-				MessageBox.Show($"При сохранении БД возникла ошибка\n{ex.Message}", "Ошибка сохранения БД", MessageBoxButtons.OK, MessageBoxIcon.Error);
-				res = false;
-			}
-			return res;
-		}
-		*/
-
-
 }
