+++ Lesson08/Task03/Question.cs
@@ -1,0 +1,22 @@
+﻿using System;
+
+namespace Task03
+{
+	[Serializable]
+	public class Question
+	{
+		public string text;       // Текст вопроса
+		public bool trueFalse;// Правда или нет
+													// Здесь мы нарушаем правила инкапсуляции и эти поля нужно было бы реализовать через открытые свойства, но для упрощения примера оставим так
+													// Вам же предлагается сделать поля закрытыми и реализовать открытые свойства Text и TrueFalse
+													// Для сериализации должен быть пустой конструктор.
+		public Question()
+		{
+		}
+		public Question(string text, bool trueFalse)
+		{
+			this.text = text;
+			this.trueFalse = trueFalse;
+		}
+	}
+}
